Split code into seperate files
